# docs/admin/swagger.yaml
basePath: /api/admin
definitions:
  base_info.Success:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      msg:
        example: err msg
        type: string
      reason:
        description: 错误原因，定义为业务判定错误码
        example: success
        type: string
    type: object
  base_info.Swagger400Resp:
    properties:
      code:
        example: 400
        type: integer
      msg:
        example: err msg
        type: string
      reason:
        example: err_reason
        type: string
    type: object
  base_info.Swagger401Resp:
    properties:
      code:
        example: 401
        type: integer
      msg:
        example: err msg
        type: string
      reason:
        example: err_reason
        type: string
    type: object
  base_info.Swagger500Resp:
    properties:
      code:
        example: 500
        type: integer
      msg:
        example: err msg
        type: string
      reason:
        example: err_reason
        type: string
    type: object
...
  commands.AssignTenantPermissionsCommand:
    properties:
      permissionIds:
        items:
          type: integer
        type: array
      tenantId:
        type: string
    required:
    - permissionIds
    - tenantId
    type: object
...
  commands.CreatePermissionsCommand:
    properties:
      code:
        type: string
      description:
        maxLength: 200
        type: string
      icon:
        type: string
      localize:
        type: string
      name:
        type: string
      parentId:
        type: integer
      path:
        type: string
      properties:
        type: string
      resources:
        items:
          $ref: '#/definitions/commands.CreatePermissionsResourceCommand'
        type: array
      sequence:
        minimum: 0
        type: integer
      type:
        enum:
        - 1
        - 2
        - 3
        type: integer
    required:
    - code
    - name
    - type
    type: object
  commands.CreatePermissionsResourceCommand:
    properties:
      method:
        enum:
        - GET
        - POST
        - PUT
        - DELETE
        type: string
      path:
        type: string
    required:
    - method
    - path
    type: object
...
  commands.UpdatePermissionsCommand:
    properties:
      description:
        maxLength: 200
        type: string
      icon:
        type: string
      id:
        type: integer
      localize:
        type: string
      name:
        type: string
      parentId:
        type: integer
      path:
        type: string
      properties:
        type: string
      resources:
        items:
          $ref: '#/definitions/commands.CreatePermissionsResourceCommand'
        type: array
      sequence:
        minimum: 0
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
      type:
        enum:
        - 1
        - 2
        - 3
        type: integer
    required:
    - id
    type: object
...
  dto.OperationLogDto:
    properties:
      action:
        description: 操作类型
        type: string
      body:
        description: 请求体
        type: string
      createdAt:
        description: 创建时间
        type: integer
      duration:
        description: 执行时长(ms)
        type: integer
      error:
        description: 错误信息
        type: string
      id:
        type: integer
      ip:
        description: 请求IP
        type: string
      method:
        description: 请求方法
        type: string
      module:
        description: 模块名称
        type: string
      path:
        description: 请求路径
        type: string
      query:
        description: 查询参数
        type: string
      status:
        description: 响应状态码
        type: integer
      tenant_id:
        description: 租户ID
        type: string
      user_agent:
        description: 用户代理
        type: string
      user_id:
        description: 操作人ID
        type: string
      username:
        description: 操作人用户名
        type: string
    type: object
  dto.PermissionsDto:
    properties:
      children:
        description: 子权限
        items:
          $ref: '#/definitions/dto.PermissionsDto'
        type: array
      code:
        description: 编码
        type: string
      createdAt:
        description: 创建时间
        type: integer
      description:
        description: 描述
        type: string
      icon:
        description: 图标
        type: string
      id:
        description: ID
        type: integer
      localize:
        description: 本地化
        type: string
      name:
        description: 名称
        type: string
      parentId:
        description: 父级ID
        type: integer
      parentPath:
        description: 父级路径
        type: string
      path:
        description: 路径
        type: string
      properties:
        description: 属性
        type: string
      resources:
        description: 资源列表
        items:
          $ref: '#/definitions/dto.PermissionsResourceDto'
        type: array
      sequence:
        description: 排序
        type: integer
      status:
        description: 状态
        type: integer
      type:
        description: 类型
        type: integer
      updatedAt:
        description: 更新时间
        type: integer
    type: object
  dto.PermissionsResourceDto:
    properties:
      method:
        description: HTTP方法
        type: string
      path:
        description: 资源路径
        type: string
    type: object
  dto.PermissionsTreeDto:
    properties:
      children:
        items:
          $ref: '#/definitions/dto.PermissionsTreeDto'
        type: array
      code:
        description: 编码
        type: string
      description:
        description: 描述
        type: string
      icon:
        description: 图标
        type: string
      id:
        description: ID
        type: integer
      localize:
        description: 本地化
        type: string
      name:
        description: 名称
        type: string
      parent_path:
        description: 父级路径
        type: string
      parentId:
        description: 父级ID
        type: integer
      path:
        description: 路径
        type: string
      properties:
        description: 属性
        type: string
      sequence:
        description: 排序
        type: integer
      status:
        description: 状态
        type: integer
      type:
        description: 类型
        type: integer
    type: object
  dto.PermissionsTreeResult:
    properties:
      ids:
        items:
          type: integer
        type: array
      tree:
        items:
          $ref: '#/definitions/dto.PermissionsTreeDto'
        type: array
    type: object
...
host: localhost:8888
info:
  contact:
    name: go-server-template-admin
  description: This is a demo using go-server-template-admin.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: ares-ddd-admin
  version: "1.0"
paths:
...
  /v1/oplog/list:
    get:
      consumes:
      - application/json
      description: 查询操作日志列表
      parameters:
      - description: 租户ID
        in: query
        name: tenant_id
        required: true
        type: string
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 模块
        in: query
        name: module
        type: string
      - description: 操作类型
        in: query
        name: action
        type: string
      - description: 开始时间
        in: query
        name: start_time
        type: integer
      - description: 结束时间
        in: query
        name: end_time
        type: integer
      - description: 页码
        in: query
        name: current
        type: integer
      - description: 每页大小
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.OperationLogDto'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 查询操作日志列表
      tags:
      - 操作日志
...
  /v1/sys/permissions:
    get:
      description: 获取权限列表
      operationId: PermissionsList
      parameters:
      - description: 编码
        in: query
        name: code
        type: string
      - description: 页码，从1开始
        in: query
        name: current
        type: integer
      - description: 名称
        in: query
        name: name
        type: string
      - description: 页码大小，最大100
        in: query
        name: size
        type: integer
      - description: 角色状态（禁用、启用）
        in: query
        name: status
        type: integer
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.PermissionsDto'
                  type: array
              type: object
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取权限列表
      tags:
      - 系统权限
    post:
      description: 添加权限
      operationId: AddPermissions
      parameters:
      - description: 属性说明请在对应model中查看
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/commands.CreatePermissionsCommand'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 添加权限
      tags:
      - 系统权限
    put:
      consumes:
      - application/json
      description: 更新权限信息，包括基本信息和资源关联
      operationId: UpdatePermissions
      parameters:
      - description: 权限更新信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/commands.UpdatePermissionsCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 更新权限
      tags:
      - 系统权限
  /v1/sys/permissions/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定ID的权限
      operationId: DeletePermissions
      parameters:
      - description: 权限ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 删除权限
      tags:
      - 系统权限
  /v1/sys/permissions/enabled:
    get:
      consumes:
      - application/json
      description: 获取系统中所有启用状态的权限列表
      operationId: GetAllEnabledPermissions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.PermissionsDto'
                  type: array
              type: object
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取所有启用状态的权限
      tags:
      - 系统权限
  /v1/sys/permissions/simple/tree:
    get:
      consumes:
      - application/json
      description: 获取简化的权限树结构
      operationId: GetPermissionsSimpleTree
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  $ref: '#/definitions/dto.PermissionsTreeResult'
              type: object
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取简化的权限树结构
      tags:
      - 系统权限
  /v1/sys/permissions/tree:
    get:
      consumes:
      - application/json
      description: 获取权限树形结构
      operationId: GetPermissionsTree
      parameters:
      - description: 权限类型
        in: query
        name: type
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.PermissionsDto'
                  type: array
              type: object
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取权限树
      tags:
      - 系统权限
...
  /v1/sys/tenant/permissions:
    put:
      consumes:
      - application/json
      description: 为指定租户分配权限
      operationId: AssignTenantPermissions
      parameters:
      - description: 权限分配信息
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/commands.AssignTenantPermissionsCommand'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/base_info.Success'
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 分配权限给租户
      tags:
      - 系统租户
  /v1/sys/tenant/permissions/:id:
    get:
      consumes:
      - application/json
      description: 获取指定租户的权限列表
      operationId: GetTenantPermissions
      parameters:
      - description: 租户ID
        in: query
        name: tenant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/base_info.Success'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.PermissionsDto'
                  type: array
              type: object
        "400":
          description: code为400 参数输入错误
          schema:
            $ref: '#/definitions/base_info.Swagger400Resp'
        "401":
          description: code为401 token未带上
          schema:
            $ref: '#/definitions/base_info.Swagger401Resp'
        "500":
          description: code为500 服务端内部错误
          schema:
            $ref: '#/definitions/base_info.Swagger500Resp'
      summary: 获取租户权限
      tags:
      - 系统租户
...
schemes:
- http
swagger: "2.0"